Reapply "issues-map: relocate domain implementation to cnos.issues package boundary"
This reverts commit f9707a9d4c0301de9067b15bc17b319e27adfc04.
