+++ src/packages/cnos.issues/commands/issues-map/go.mod
@@ -1,0 +1,3 @@
+module github.com/usurobor/cnos/packages/cnos.issues/commands/issues-map
+
+go 1.24
