+++ src/packages/cnos.issues/commands/issues-map/templates/board.template.html
